# api/openapi/identity-service.v1.yaml
openapi: 3.0.3
info:
  title: identity-service API
  version: 1.0.0
  description: |
    Contract for the go-ory-auth-example identity-service (ADR-0010).
    Human-readable guide: docs/api/identity-service-api.md.

    Credential is bound to plane:
    - `/v1/*` accepts only `Authorization: Bearer <kratos session token>` (mobile).
    - `/admin/v1/*` accepts only the Kratos session cookie (admin web).

    Errors use RFC 9457 Problem Details (`application/problem+json`) with a stable `code`.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /internal/hooks/kratos/after-registration:
    post:
      tags: [hooks]
      operationId: kratosAfterRegistration
      summary: Kratos after-registration webhook (profile provisioning)
      security: [{ webhookKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KratosHookPayload" }
      responses:
        "204": { description: Processed (idempotent) }
        "401": { description: Bad key }
  /internal/hooks/kratos/after-login:
    post:
      tags: [hooks]
      operationId: kratosAfterLogin
      summary: Kratos after-login population guard
      description: |
        Interrupting population guard (ADR-0004). Returns 204 to allow, or 403
        with a Kratos-formatted message body to interrupt the login.
      security: [{ webhookKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KratosHookPayload" }
      responses:
        "204": { description: Allowed }
        "401": { description: Bad key }
        "403": { description: Login interrupted (Kratos webhook message format) }
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      description: Kratos session token from a native (API) flow.
    cookieSession:
      type: apiKey
      in: cookie
      name: ory_kratos_session
    webhookKey:
      type: apiKey
      in: header
      name: Authorization
...
  schemas:
...
    KratosHookPayload:
      type: object
      required: [identity_id, schema_id, flow_type]
      properties:
        identity_id: { type: string, format: uuid }
        schema_id: { type: string }
        flow_type: { type: string, enum: [api, browser] }